third_party/blink/web_tests/fast/dom/shadow/getelementbyid-in-orphan-expected.txt

Tests to ensure ShadowRoot.getElementById works even if a ShadowRoot is orphan.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS shadowRoot.getElementById("child1") is shadowChild1
PASS shadowRoot.getElementById("child2") is shadowChild2
PASS shadowRoot.getElementById("child3") is null
PASS shadowRoot.getElementById("child4") is null
PASS document.getElementById("child1") is null
PASS document.getElementById("child2") is null
PASS document.getElementById("child3") is null
PASS document.getElementById("child4") is null


Insert elements having the same id
PASS shadowRoot.getElementById("child2") is shadowChild2
PASS shadowRoot.getElementById("child2") is shadowChild2_3


Make the host in document
PASS shadowRoot.getElementById("child1") is shadowChild1
PASS shadowRoot.getElementById("child2") is shadowChild2_3
PASS shadowRoot.getElementById("child3") is null
PASS shadowRoot.getElementById("child4") is null
PASS document.getElementById("child1") is null
PASS document.getElementById("child2") is null
PASS document.getElementById("child3") is child3
PASS document.getElementById("child4") is child4


Add a child and make the host not in document
PASS shadowRoot.getElementById("child5") is shadowChild5


PASS successfullyParsed is true

TEST COMPLETE
